requirements.txt
langchain==0.0.313
python-dotenv==1.0.0
streamlit==1.26.0
openai==0.27.9
pymupdf==1.22.5
faiss-cpu==1.7.4